MM-187 feat: 특정 월 유저의 참여 신청한 일정 조회 URI 변경 (#219)
