+++ src/main/java/com/dev/moim/domain/moim/service/impl/CalenderQueryServiceImpl.java
@@ -9,7 +9,6 @@ import com.dev.moim.domain.moim.entity.UserPlan;
 import com.dev.moim.domain.moim.entity.enums.JoinStatus;
 import com.dev.moim.domain.moim.repository.*;
 import com.dev.moim.domain.moim.service.CalenderQueryService;
-import com.dev.moim.domain.user.dto.UserPlanDTO;
 import com.dev.moim.global.error.handler.MoimException;
 import com.dev.moim.global.error.handler.PlanException;
 import lombok.RequiredArgsConstructor;
@@ -41,21 +40,6 @@ public class CalenderQueryServiceImpl implements CalenderQueryService {
     private final UserPlanRepository userPlanRepository;
     private final ScheduleRepository scheduleRepository;
     private final UserMoimRepository userMoimRepository;
-
-    @Override
-    public PlanMonthListDTO<List<UserPlanDTO>> getUserPlans(User user, int year, int month) {
-        YearMonth yearMonth = YearMonth.of(year, month);
-        LocalDateTime startDate = yearMonth.atDay(1).atStartOfDay();
-        LocalDateTime endDate = yearMonth.atEndOfMonth().atTime(23, 59, 59);
-
-        List<UserPlan> userPlanList = userPlanRepository.findByUserIdAndPlanDateBetween(user.getId(), startDate, endDate);
-
-        Map<Integer, List<UserPlanDTO>> planListByDay = userPlanList.stream()
-                .map(userPlan -> UserPlanDTO.toUserMoimPlan(userPlan.getPlan()))
-                .collect(Collectors.groupingBy(dto -> dto.time().getDayOfMonth()));
-
-        return new PlanMonthListDTO<>(planListByDay);
-    }
 
     @Override
     public PlanMonthListDTO<PlanDayListDTO> getMoimPlans(User user, Long moimId, int year, int month) {
