+++ src/main/java/com/dev/moim/domain/user/service/impl/UserQueryServiceImpl.java
@@ -137,6 +137,21 @@ public class UserQueryServiceImpl implements UserQueryService {
     }
 
     @Override
+    public PlanMonthListDTO<List<UserPlanDTO>> getUserPlans(User user, int year, int month) {
+        YearMonth yearMonth = YearMonth.of(year, month);
+        LocalDateTime startDate = yearMonth.atDay(1).atStartOfDay();
+        LocalDateTime endDate = yearMonth.atEndOfMonth().atTime(23, 59, 59);
+
+        List<UserPlan> userPlanList = userPlanRepository.findByUserIdAndPlanDateBetween(user.getId(), startDate, endDate);
+
+        Map<Integer, List<UserPlanDTO>> planListByDay = userPlanList.stream()
+                .map(userPlan -> UserPlanDTO.toUserMoimPlan(userPlan.getPlan()))
+                .collect(Collectors.groupingBy(dto -> dto.time().getDayOfMonth()));
+
+        return new PlanMonthListDTO<>(planListByDay);
+    }
+
+    @Override
     public PlanMonthListDTO<List<UserPlanDTO>> getUserMonthlyPlans(User user, int year, int month) {
         YearMonth yearMonth = YearMonth.of(year, month);
         LocalDateTime startDate = yearMonth.atDay(1).atStartOfDay();
